080523-1

+++ ft_bzero.c
@@ -6,14 +6,16 @@
 /*   By: andmart2 <andmart2@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/02 16:17:20 by andmart2          #+#    #+#             */
-/*   Updated: 2023/05/04 15:26:30 by andmart2         ###   ########.fr       */
+/*   Updated: 2023/05/08 18:43:56 by andmart2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void ft_bzero(void *s, unsigned int n)
+#include "libft.h"
+
+void ft_bzero(void *s, size_t n)
 {
 	unsigned char *str = (unsigned char *)s;
-	unsigned  int i;
+	size_t i;
 
 	i = 0;
 	while ( i < n)
